Added sortier-ignore-next-line support (#964)
I've added two comments that are similar to the eslint-disable-next-line ability available in eslint

    sortier-ignore-next-line - Disables sorting on all AST nodes that start and end on the next line
    sortier-ignore-nodes - Disables sorting on all AST nodes that start on the next line

The test files give a decent array of scenarios where it maybe useful. I'm not considering these to be fully flushed out and there are still situations where there still isn't a way to disable ordering (ex. class ordering or import ordering) but it's at least a good start.